1. Error handling image uploads on invalid inputs
2. Formatting update and delete links on widescreen

3. Formatting update forms for widescreen

+++ controllers/computerpartController.js
@@ -275,15 +275,25 @@ exports.computerpart_update_post = [
         _id: req.params.id,
       });
 
-      if (req.file) {
+      if (req.file && errors.isEmpty()) {
+        console.log('FILE', req.file)
         component.fileName = req.file.filename;
         fs.unlink(`public/images/${req.body.fileName}`, (err) => {
           if (err) console.log(err);
           console.log(req.body.fileName, "was deleted");
         });
+      } else if(req.body.fileName && req.body.fileName !='null' && req.body.fileName !='undefined'){
+        component.fileName = req.body.fileName;
       }
 
       if (!errors.isEmpty()) {
+        if(req.file){
+          fs.unlink(`public/images/${req.file.filename}`, (err) => {
+            if (err) console.log(err);
+            console.log(req.file.filename, "was deleted");
+          });
+        }
+        
         async.parallel(
           {
             categories: function (callback) {
